+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `alwaysLeafCategoryAuction` settings parameter to run auctions only over leaf categories
 - Added `activateDebugSponsoredTags` settings parameter to display the (Ad) tag next to sponsored products
 
+## [2.1.1] - 2025-03-11
+
+### Added
+
+- Fix several issues
+
 ## [2.1.0] - 2025-03-11
 
 ### Added